Add super-save layer that auto-saves buffers
This layer is inspired by Bozhidar Batsovs
super-save (https://github.com/bbatsov/super-save), but is adapted to
work with Spacemacs. Using the original package proved to be impractical
for integrating with Spacemacs, and since it is not a large package, the
functionality was reimplemented.
When you enable this layer, switching windows or Spacemacs losing focus
causes an automatic save to the current buffer.